🔀 MB Merge to develop

See merge request s11-webmobile4-sub2/S11P12D110!88

+++ NativeNavs-Android/NativeNavs/app/src/main/java/com/circus/nativenavs/data/ProfileUserDto.kt
@@ -11,7 +11,7 @@ data class ProfileUserDto(
     val korean: Boolean,
     val name: String,
     val nation: String,
-    val navReviewAverage: Int,
+    val navReviewAverage: Double,
     val navReviewCount: Int,
     var nickname: String,
     var phone: String,
